+++ libs/computer/typescript/src/interface/types.ts
