+++ libs/computer/typescript/src/types.ts
